requirements.txt
Flask==1.1.2
passlib==1.7.2
WTForms==2.3.1
Flask_SQLAlchemy==2.4.3
Flask_WTF==0.14.3
python-dotenv==0.14.0
